Wine_trade/apps/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
pyjwt>=2.8.0
cryptography>=41.0.0
httpx>=0.25.0
requests>=2.31.0
slowapi>=0.1.9
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
sentence-transformers>=2.2.0
numpy>=1.24.0
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
joblib>=1.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-snapshot>=0.9.0

